refactor: 폴더 구조 변경 및 리펙토링
